+++ kopy-args/jvm/main/kotlin/com/javiersc/kotlin/kopy/args/KopyArgument.jvm.kt
@@ -1,0 +1,5 @@
+package com.javiersc.kotlin.kopy.args
+
+import java.io.Serializable
+
+public actual interface KopyArgument : Serializable